+++ lib/include/bave/io/file_io.hpp
@@ -4,11 +4,21 @@
 #include <vector>
 
 namespace bave::file {
-/// \brief Find a super directory that contains a given file/directory pattern.
-/// \param base Directory to start search from.
+/// \brief Find a super path to a given file/directory pattern.
+/// \param base Directory to start search from. The working directory is always searched.
 /// \param patterns Comma-separated list of file patterns to search for.
-/// \returns Path to super directory if found, else empty string.
-[[nodiscard]] auto find_super_dir(std::string_view base, std::string_view patterns) -> std::string;
+/// \returns Path to file/directory if found, else empty string.
+[[nodiscard]] auto upfind(std::string_view base, std::string_view patterns) -> std::string;
+/// \brief Find a super path to a given directory pattern.
+/// \param base Directory to start search from. The working directory is always searched.
+/// \param patterns Comma-separated list of directory name patterns to search for.
+/// \returns Path to directory if found, else empty string.
+[[nodiscard]] auto upfind_dir(std::string_view base, std::string_view patterns) -> std::string;
+/// \brief Find the parent directory of a super directory containing a given file.
+/// \param base Directory to start search from. The working directory is always searched.
+/// \param filename Filename to search for.
+/// \returns Path to parent directory if found, else empty string.
+[[nodiscard]] auto upfind_parent(std::string_view base, std::string_view filename) -> std::string;
 
 /// \brief Check if a file exists.
 /// \param path Path to check for.
@@ -25,4 +35,10 @@ namespace bave::file {
 /// \param path Path to read from.
 /// \returns true on success.
 [[nodiscard]] auto read_string(std::string& out, CString path) -> bool;
+
+/// \brief Create a URI relative to the assets path. Only relevant for desktop platforms.
+/// \param assets_dir Path to assets directory.
+/// \param full_path Full path to convert to URI.
+/// \returns URI relative to the assets path.
+[[nodiscard]] auto make_uri(std::string_view assets_dir, std::string_view full_path) -> std::string;
 } // namespace bave::file