+++ lib/src/app.cpp
@@ -3,30 +3,11 @@
 #include <bave/core/error.hpp>
 #include <bave/driver.hpp>
 #include <capo/error_handler.hpp>
-#include <filesystem>
 
 namespace bave {
-namespace fs = std::filesystem;
-
-struct App::ZipFs {
-	explicit ZipFs() : m_instance(PHYSFS_init(nullptr)) {
-		if (m_instance == 0) { m_log.warn("failed to initialize ZipFs"); }
-	}
-
-  private:
-	struct Deleter {
-		void operator()(int /*i*/) const noexcept { PHYSFS_deinit(); }
-	};
-
-	Logger m_log{"ZipFs"};
-	ScopedResource<int, Deleter> m_instance{};
-};
-
-void App::ZipFsDeleter::operator()(Ptr<ZipFs> ptr) const noexcept { std::default_delete<ZipFs>{}(ptr); }
-
 App::App(std::string tag)
 	: m_log{std::move(tag)}, m_bootloader([](App& app) { return std::make_unique<Driver>(app); }), m_audio_device(std::make_unique<AudioDevice>()),
-	  m_audio_streamer(std::make_unique<AudioStreamer>(*m_audio_device)), m_zip_fs(new ZipFs{}) {
+	  m_audio_streamer(std::make_unique<AudioStreamer>(*m_audio_device)) {
 	log::get_thread_id(); // set thread 0
 
 	for (int i = 0; i < static_cast<int>(m_gamepads.size()); ++i) { m_gamepads.at(static_cast<GamepadId>(i)).id = GamepadId{i}; }
@@ -91,14 +72,7 @@ auto App::load_shader(std::string_view vertex, std::string_view fragment) const 
 	return Shader{&get_renderer(), vert, frag};
 }
 
-void App::add_data_loader(std::unique_ptr<IDataLoader> loader, int priority) { m_data_store->add_loader(std::move(loader), priority); }
-
-auto App::make_uri(std::string_view const full_path) const -> std::string {
-	if (full_path.empty()) { return {}; }
-	auto const assets_path = get_assets_path();
-	if (assets_path.empty()) { return std::string{full_path}; }
-	return fs::path{full_path}.lexically_relative(assets_path).generic_string();
-}
+void App::set_data_loader(std::unique_ptr<IDataLoader> loader) { m_data_store->set_loader(std::move(loader)); }
 
 auto App::get_features() const -> FeatureFlags {
 	auto ret = do_get_native_features();