+++ lib/src/data_store.cpp
@@ -1,6 +1,5 @@
 #include <bave/data_store.hpp>
 #include <djson/json.hpp>
-#include <algorithm>
 #include <filesystem>
 
 namespace bave {
@@ -17,37 +16,28 @@ auto DataStore::as_string_view(std::span<std::byte const> bytes) -> std::string_
 	return {reinterpret_cast<char const*>(bytes.data()), bytes.size()}; // NOLINT(cppcoreguidelines-pro-type-reinterpret-cast)
 }
 
-void DataStore::add_loader(std::unique_ptr<IDataLoader> loader, int const priority) {
+void DataStore::set_loader(std::unique_ptr<IDataLoader> loader) {
 	if (!loader) { return; }
-	auto const it = std::lower_bound(m_loaders.begin(), m_loaders.end(), priority, [](Entry const& a, int const priority) { return a.priority < priority; });
-	auto entry = Entry{.loader = std::move(loader), .priority = priority};
-	m_loaders.insert(it, std::move(entry));
+	m_loader = std::move(loader);
 }
 
 auto DataStore::exists(std::string_view uri) const -> bool {
-	if (uri.empty()) { return false; }
-	for (auto const& [loader, _] : m_loaders) {
-		if (loader->exists(uri)) { return true; }
-	}
-	return false;
+	if (uri.empty() || !m_loader) { return false; }
+	return m_loader->exists(uri);
 }
 
 auto DataStore::read_bytes(std::string_view uri) const -> std::vector<std::byte> {
-	if (uri.empty()) { return {}; }
+	if (uri.empty() || !m_loader) { return {}; }
 	auto ret = std::vector<std::byte>{};
-	for (auto const& [loader, _] : m_loaders) {
-		if (loader->read_bytes(ret, uri)) { return ret; }
-	}
+	if (m_loader->read_bytes(ret, uri)) { return ret; }
 	m_log.warn("failed to load data at: '{}'", uri);
 	return {};
 }
 
 auto DataStore::read_string(std::string_view uri) const -> std::string {
-	if (uri.empty()) { return {}; }
+	if (uri.empty() || !m_loader) { return {}; }
 	auto ret = std::string{};
-	for (auto const& [loader, _] : m_loaders) {
-		if (loader->read_string(ret, uri)) { return ret; }
-	}
+	if (m_loader->read_string(ret, uri)) { return ret; }
 	m_log.warn("failed to load data at: '{}'", uri);
 	return {};
 }