+++ lib/src/io/zip_io_reinit.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <bave/core/c_string.hpp>
+
+namespace bave::zip {
+auto reinit(CString arg0) -> bool;
+}